% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facerec_init.R
\name{facerec_init}
\alias{facerec_init}
\title{authorization}
\usage{
facerec_init()
}
\value{
nothing.
}
\description{
Initializes the authorization credentials for the Kairos Face Recognition API.
Needs to be called before using any other functions of \code{facerec} and requires \code{kairos_id} and \code{kairos_key} as environment variables.
}
\examples{
\dontrun{
Sys.setenv(kairos_id = "Your Kairos API id")
Sys.setenv(kairos_key = "Your Kairos API key")

facerec_init()
}




}
