% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_galleries.R
\name{get_galleries}
\alias{get_galleries}
\title{list galleries}
\usage{
get_galleries()
}
\value{
A vector of gallery id's.
}
\description{
Returns identifiers for all galleries associated with a 'Kairos' application.
}
\examples{
\dontrun{
# enroll
finn_image <- 'https://upload.wikimedia.org/wikipedia/en/2/2a/Finn-Force_Awakens_\%282015\%29.png'
first_gallery <- enroll(image = finn_image, subject_id = 'finn', gallery = 'starwars')
second_gallery <- enroll(image = finn_image, subject_id = 'finn', gallery = 'more_starwars')

# get_galleries
get_galleries()
}


}
