% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gallery_subjects.R
\name{get_gallery_subjects}
\alias{get_gallery_subjects}
\title{get gallery subjects}
\usage{
get_gallery_subjects(gallery)
}
\arguments{
\item{gallery}{The gallery in which the subjects are enrolled.}
}
\value{
A vector of subject id's
}
\description{
Returns all subject id's associated with a gallery.
}
\examples{
\dontrun{
# enroll
finn_image <- 'https://upload.wikimedia.org/wikipedia/en/2/2a/Finn-Force_Awakens_\%282015\%29.png'
finn_enroll <- enroll(image = finn_image, subject_id = 'finn', gallery = 'starwars')

# view subjects
get_gallery_subjects(gallery = 'starwars')
}


}
