% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_gallery.R
\name{remove_gallery}
\alias{remove_gallery}
\title{remove gallery}
\usage{
remove_gallery(gallery)
}
\arguments{
\item{gallery}{The name of the gallery to be removed.}
}
\value{
nothing.
}
\description{
Removes a gallery and all included subjects.
}
\examples{
\donttest{


facerec_init()

# enroll
finn_image <- 'https://upload.wikimedia.org/wikipedia/en/2/2a/Finn-Force_Awakens_\%282015\%29.png'
finn_enroll <- enroll(image = finn_image, subject_id = 'finn', gallery = 'starwars')

# remove gallery
remove_gallery(gallery = 'starwars')
}


}
