% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_subject.R
\name{remove_subject}
\alias{remove_subject}
\title{remove subject}
\usage{
remove_subject(subject_id, gallery)
}
\arguments{
\item{subject_id}{The subject id for the subject to be removed.}

\item{gallery}{The name of the gallery in which the subject is enrolled.}
}
\value{
nothing.
}
\description{
Removes a subject from a gallery.
}
\examples{
\donttest{
# enroll
finn_image <- 'https://upload.wikimedia.org/wikipedia/en/2/2a/Finn-Force_Awakens_\%282015\%29.png'
finn_enroll <- enroll(image = finn_image, subject_id = 'finn', gallery = 'starwars')

# remove subject
remove_subject(subject_id = 'finn', gallery = 'starwars')
}


}
