% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_subject.R
\name{view_subject}
\alias{view_subject}
\title{view subject}
\usage{
view_subject(subject_id, gallery)
}
\arguments{
\item{subject_id}{The subject id for which to return all face id's.}

\item{gallery}{The gallery in which the subject is enrolled.}
}
\value{
A dataframe with face id's and enrollment timestamps associated with the input subject.
}
\description{
Returns all face id's for each image enrolled for a given subject in a gallery.
}
\examples{
\donttest{
# enroll
finn_image <- 'https://upload.wikimedia.org/wikipedia/en/2/2a/Finn-Force_Awakens_\%282015\%29.png'
finn_enroll <- enroll(image = finn_image, subject_id = 'finn', gallery = 'starwars')

# view subject
view_subject(subject_id = 'finn', gallery = 'starwars')
}


}
